+++ packages/atomic/src/components.d.ts
@@ -10,6 +10,8 @@ import { Bindings } from "./utils/initialization-utils";
 import { NumberInputType } from "./components/facets/facet-number-input/number-input-type";
 import { ResultDisplayDensity, ResultDisplayImageSize, ResultDisplayLayout } from "./components/atomic-result/atomic-result-display-options";
 import { Section } from "./components/atomic-layout-section/sections";
+import { ObservableMap } from "@stencil/store";
+import { AtomicStore } from "./utils/store";
 import { i18n } from "i18next";
 import { InitializationOptions } from "./components/atomic-search-interface/atomic-search-interface";
 import { StandaloneSearchBoxData } from "./utils/local-storage-utils";
@@ -520,6 +522,9 @@ export namespace Components {
         "bindings": Bindings;
     }
     interface AtomicResult {
+        /**
+          * Classes that will be added to the result element.
+         */
         "classes": string;
         /**
           * The result content to display.
@@ -545,10 +550,15 @@ export namespace Components {
           * How large or small the visual section of results should be.  This may be overwritten if an image size is defined in the result content.
          */
         "imageSize"?: ResultDisplayImageSize;
+        "loadingFlag"?: string;
         /**
           * The result item.
          */
         "result": Result | FoldedResult;
+        /**
+          * Global state for Atomic.
+         */
+        "store"?: ObservableMap<AtomicStore>;
     }
     interface AtomicResultBadge {
         /**
@@ -641,7 +651,7 @@ export namespace Components {
         /**
           * The expected size of the image displayed in the results.
          */
-        "imageSize"?: ResultDisplayImageSize;
+        "imageSize": ResultDisplayImageSize;
         /**
           * Sets a rendering function to bypass the standard HTML template mechanism for rendering results. You can use this function while working with web frameworks that don't use plain HTML syntax, e.g., React, Angular or Vue.  Do not use this method if you integrate Atomic in a plain HTML deployment.
           * @param render
@@ -775,6 +785,10 @@ export namespace Components {
     }
     interface AtomicSearchBoxQuerySuggestions {
         /**
+          * The SVG icon to display.  - Use a value that starts with `http://`, `https://`, `./`, or `../`, to fetch and display an icon from a given location. - Use a value that starts with `assets://`, to display an icon from the Atomic package. - Use a stringified SVG to display it directly.
+         */
+        "icon"?: string;
+        /**
           * The maximum number of suggestions that will be displayed if the user has typed something into the input field.
          */
         "maxWithQuery"?: number;
@@ -784,6 +798,10 @@ export namespace Components {
         "maxWithoutQuery"?: number;
     }
     interface AtomicSearchBoxRecentQueries {
+        /**
+          * The SVG icon to display.  - Use a value that starts with `http://`, `https://`, `./`, or `../`, to fetch and display an icon from a given location. - Use a value that starts with `assets://`, to display an icon from the Atomic package. - Use a stringified SVG to display it directly.
+         */
+        "icon"?: string;
         /**
           * The maximum number of suggestions that will be displayed if the user has typed something into the input field.
          */
@@ -2062,6 +2080,9 @@ declare namespace LocalJSX {
         "bindings": Bindings;
     }
     interface AtomicResult {
+        /**
+          * Classes that will be added to the result element.
+         */
         "classes"?: string;
         /**
           * The result content to display.
@@ -2087,10 +2108,15 @@ declare namespace LocalJSX {
           * How large or small the visual section of results should be.  This may be overwritten if an image size is defined in the result content.
          */
         "imageSize"?: ResultDisplayImageSize;
+        "loadingFlag"?: string;
         /**
           * The result item.
          */
         "result": Result | FoldedResult;
+        /**
+          * Global state for Atomic.
+         */
+        "store"?: ObservableMap<AtomicStore>;
     }
     interface AtomicResultBadge {
         /**
@@ -2304,6 +2330,10 @@ declare namespace LocalJSX {
     }
     interface AtomicSearchBoxQuerySuggestions {
         /**
+          * The SVG icon to display.  - Use a value that starts with `http://`, `https://`, `./`, or `../`, to fetch and display an icon from a given location. - Use a value that starts with `assets://`, to display an icon from the Atomic package. - Use a stringified SVG to display it directly.
+         */
+        "icon"?: string;
+        /**
           * The maximum number of suggestions that will be displayed if the user has typed something into the input field.
          */
         "maxWithQuery"?: number;
@@ -2313,6 +2343,10 @@ declare namespace LocalJSX {
         "maxWithoutQuery"?: number;
     }
     interface AtomicSearchBoxRecentQueries {
+        /**
+          * The SVG icon to display.  - Use a value that starts with `http://`, `https://`, `./`, or `../`, to fetch and display an icon from a given location. - Use a value that starts with `assets://`, to display an icon from the Atomic package. - Use a stringified SVG to display it directly.
+         */
+        "icon"?: string;
         /**
           * The maximum number of suggestions that will be displayed if the user has typed something into the input field.
          */