+++ packages/atomic/src/components/result-lists/list-display-results.tsx
@@ -8,12 +8,15 @@ export const ListDisplayResults: FunctionalComponent<ResultsProps> = (
     return (
       <atomic-result
         key={props.resultListCommon.getResultId(result, props.resultListState)}
+        part="outline"
         result={result}
         engine={props.bindings.engine}
+        store={props.bindings.store}
         display={props.display}
         density={props.density}
         imageSize={props.imageSize}
         content={props.getContentOfResultTemplate(result)}
+        loadingFlag={props.resultListCommon.loadingFlag}
       ></atomic-result>
     );
   });