+++ packages/quantic/force-app/main/default/lwc/quanticDateFacet/quanticDateFacet.js
@@ -85,6 +85,14 @@ export default class QuanticDateFacet extends LightningElement {
   /** @type {boolean} */
   _isCollapsed = false;
 
+  static attributes = [
+    'facetId',
+    'field',
+    'label',
+    'numberOfValues',
+    'formattingFunction',
+  ]
+
   /** @type {DateFacetState} */
   @track state;
 
@@ -136,6 +144,7 @@ export default class QuanticDateFacet extends LightningElement {
       label: this.label,
       facetId: this.facet.state.facetId,
       format: this.formattingFunction,
+      element: this.template.host,
     });
   }