+++ packages/quantic/force-app/main/default/lwc/quanticFacet/quanticFacet.js
@@ -124,6 +124,18 @@ export default class QuanticFacet extends LightningElement {
   /** @type {boolean} */
   _isCollapsed = false;
 
+  static attributes = [
+    'facetId',
+    'field',
+    'label',
+    'numberOfValues',
+    'sortCriteria',
+    'noSearch',
+    'displayValuesAs',
+    'noFilterFacetCount',
+    'injectionDepth',
+  ]
+
   /** @type {FacetState} */
   @track state;
   
@@ -187,7 +199,8 @@ export default class QuanticFacet extends LightningElement {
     this.unsubscribe = this.facet.subscribe(() => this.updateState());
     registerToStore(this.engineId, Store.facetTypes.FACETS, {
       label: this.label,
-      facetId: this.facet.state.facetId
+      facetId: this.facet.state.facetId,
+      element: this.template.host
     });
   }